requirements.txt

AfricastalkingGateway==2.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pytz==2018.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
astroid==1.5.3
confusable-homoglyphs==3.0.0
dj-database-url==0.4.2
Django==1.11.7
django-bootstrap3==9.1.0
django-registration==2.4.1
django-tinymce==2.7.0
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pylint==1.7.4
pyperclip==1.6.0
python-decouple==3.1
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
